Created new test cases for app/javascript/components/pages/admin/forms-list/components/form-group
